requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
# awsebcli==3.20.10
billiard==4.2.0
blessed==1.20.0
boto3==1.34.29
botocore==1.34.39
celery==5.3.6
cement==2.8.2
certifi==2023.11.17
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
distro==1.9.0
Django==5.0.1
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
emoji==2.2.0
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
jmespath==1.0.1
kombu==5.3.5
numpy==1.26.3
openai==1.6.0
packaging==23.2
pandas==2.2.0
pathspec==0.10.1
prompt-toolkit==3.0.43
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
semantic-version==2.8.5
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
termcolor==1.1.0
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.64.1
transformers==4.23.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
vine==5.1.0
wcwidth==0.1.9
whitenoise==6.6.0
python-dotenv
mailjet_rest
cloudinary
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
